# docs/openapi/smartranking-openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "Users_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/profiles": {
      "post": {
        "operationId": "Users_upsertProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/profiles/self": {
      "post": {
        "operationId": "Users_upsertSelfProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSelfProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/profiles/{userId}": {
      "get": {
        "operationId": "Users_getProfile",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Tenant": []
          },
          {
            "SessionCookie": []
          }
        ],
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "info": {
    "title": "SmartRanking API",
    "description": "Multi-tenant tennis ranking API with Better Auth, RBAC, and MongoDB.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session-token",
        "description": "Better Auth session token set via /api/auth"
      },
      "Tenant": {
        "type": "apiKey",
        "in": "header",
        "name": "x-tenant-id",
        "description": "Tenant/club identifier. Required for all protected multi-tenant endpoints."
      }
    },
    "schemas": {
...
      "CreateUserProfileDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSelfProfileDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}